requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.6,<1.0.0
pydantic>=2.11.0,<3.0.0
numpy>=2.0.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
google-generativeai
pymongo
motor
python-dotenv
